add how to deal with questions < 50 and others
On run_file.py -> final_ans() : added how to deal with a question less
than 50 by substracting number of answer given to 50 then the rest will
be decleared -1(if correct 1 incorrect 0).
Added registration in views.py and admin page.
Twicked/added on the utlis.py -> show_answer() grade[x]==0(incorrect).
Added only how to deal with front qrcode by reading the qrcode it will
have a tuple like structure that the first one is a 'front' if its a
front page(not yet on the back) in process_file.py -> upload_file()
Edited the process_answer_1.py -> func_choose on how to deal with the
correct, incorrect, disquaified and no answer.